requirements.txt
google-api-python-client
beautifulsoup4
requests
webkit2png
Pillow
Flask
slacker
selenium
Flask-Testing
flask-sqlalchemy